+++ src/main/kotlin/tech/gdragon/db/Database.kt
@@ -1,19 +1,36 @@
 package tech.gdragon.db
 
-import org.jetbrains.exposed.sql.SchemaUtils
-import org.jetbrains.exposed.sql.transactions.transaction
+import mu.KotlinLogging
+import org.flywaydb.core.Flyway
 import org.joda.time.DateTimeZone
-import tech.gdragon.db.table.Tables
 import org.jetbrains.exposed.sql.Database as ExposedDatabase
 
 interface Database
 
 class EmbeddedDatabase(dataDirectory: String) : Database {
+  val logger = KotlinLogging.logger { }
+
   init {
-    ExposedDatabase.connect("jdbc:h2:file:$dataDirectory/settings.db", "org.h2.Driver")
-    transaction {
-      SchemaUtils.create(*Tables.allTables)
+    val url = "jdbc:h2:file:$dataDirectory/settings.db"
+    logger.info {
+      "Starting database migration: $url"
     }
+    val flyway = Flyway.configure()
+      .dataSource(url, "", "")
+      .baselineOnMigrate(true)
+      .locations("h2")
+      .load()
+
+    flyway
+      .migrate()
+      .migrations
+      .forEach { migration ->
+        logger.info {
+          "Performed migration step: ${migration.description}"
+        }
+      }
+
+    ExposedDatabase.connect(url, "org.h2.Driver")
   }
 }