+++ src/main/kotlin/tech/gdragon/discord/Bot.kt
@@ -53,7 +53,7 @@ class Bot : KoinComponent {
       // create shard manager
       shardManager = DefaultShardManagerBuilder
         .create(token, GatewayIntent.GUILD_MESSAGES, GatewayIntent.GUILD_VOICE_STATES)
-        .disableCache(CacheFlag.ACTIVITY, CacheFlag.EMOTE, CacheFlag.CLIENT_STATUS)
+        .disableCache(CacheFlag.ACTIVITY, CacheFlag.EMOTE, CacheFlag.CLIENT_STATUS, CacheFlag.ONLINE_STATUS)
         .setChunkingFilter(ChunkingFilter.NONE)
         .setMemberCachePolicy(MemberCachePolicy.VOICE)
         .addEventListeners(EventListener(), SystemEventListener())